Fix: Rename armv8r_cp15.h.h to armv8r_cp15.h
Correct duplicated file extension in r-profile directory to match other
architectures.
